refactor: MainMobileHeader -> MobileHeader로 변경
