+++ frontend/src/pages/MainPage/components/MobileHeader/MobileHeader.styles.ts
